extract ParameterTypeFromDataProviderResolver
