--- rules-tests/TypeDeclarationDocblocks/Rector/ClassMethod/AddParamArrayDocblockFromDataProviderRector/Fixture/some_test_with_data_provider.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Rector\Tests\TypeDeclarationDocblocks\Rector\ClassMethod\AddParamArrayDocblockFromDataProviderRector\Fixture;
-
-use PHPUnit\Framework\Attributes\DataProvider;
-use PHPUnit\Framework\TestCase;
-
-final class SomeTestWithDataProvider extends TestCase
-{
-    #[DataProvider('provideData')]
-    public function test(array $names): void
-    {
-    }
-
-    public static function provideData()
-    {
-        yield [['Tom', 'John']];
-    }
-}
-
-?>
